+++ src/main/java/com/maptracker/issuemap/common/jwt/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.maptracker.issuemap.domain.user.dto;
+package com.maptracker.issuemap.common.jwt;
 
 import com.maptracker.issuemap.domain.user.entity.User;
 import java.util.ArrayList;
