--- src/main/java/com/maptracker/issuemap/domain/user/service/CustomUserDetailService.java
@@ -1,25 +1,0 @@
-package com.maptracker.issuemap.domain.user.service;
-
-import com.maptracker.issuemap.common.error.UserErrorCode;
-import com.maptracker.issuemap.domain.user.dto.CustomUserDetails;
-import com.maptracker.issuemap.domain.user.exception.UserException;
-import com.maptracker.issuemap.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findByEmail(username)
-                .map(CustomUserDetails::new)
-                .orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
-    }
-}
